Components

Components

This package has implemented some basic components which inherit from the abstract types defined at Interfaces.

Components

Environments

Spaces

Agents

Buffers

Learners

Approximators

Environment Models

Policies

Action Selectors

Environments

Ju.SimpleRandomWalkEnv — Type.

SimpleRandomWalkEnv

A simple random walk environment for tutorial.

Example

julia> env = SimpleRandomWalkEnv()
SimpleRandomWalkEnv(7, 3, 3, [-1, 1])

julia> env(1)
(observation = 2, reward = 0.0, isdone = false)

julia> reset!(env)
(observation = 3, isdone = false)

source

Spaces

Ju.ContinuousSpace — Type.

struct ContinuousSpace{T<:Number} <: AbstractContinuousSpace
    low::T
    high::T
end

The lower bound and upper bound are specifed by low and high.

source

Ju.DiscreteSpace — Type.

struct DiscreteSpace <: AbstractDiscreteSpace
    n::Int
end

The elements in a DiscreteSpace is 1:n

source

Ju.MultiContinuousSpace — Type.

MultiContinuousSpace(low::Number, high::Number, size::Tuple{Vararg{Int}})
MultiContinuousSpace(low::Array{<:Number}, high::Array{<:Number})

Examples

MultiContinuousSpace(-1, 1, (2,3))
MultiContinuousSpace([0, 0, 0], [1, 2, 3])

source

Ju.MultiDiscreteSpace — Type.

struct MultiDiscreteSpace{N} <:AbstractDiscreteSpace
    counts::Array{Int, N}
end

The element in MultiDiscreteSpace{N} is a multi-dimension array. The number of each dimension is specified by counts.

source

Agents

Ju.Agent — Type.

Agent{Tl<:AbstractLearner, Tb<:AbstractTurnBuffer, Tpp<:Function} <: AbstractAgent 
Agent(learner::Tl, buffer::Tb, preprocessor::Tpp=identity, role=:anonymous) where {Tl<:AbstractLearner, Tb<:AbstractTurnBuffer, Tpp<:Function}

A preprocessor is just a normal function. It transforms the observation from an environment to the internal state, which is then stored in the buffer. role is a Symbol. Usually it is used in multi-agents environment to distinction different agents.

See also: AbstractLearner, AbstractTurnBuffer

source

Ju.Agent — Method.

(agent::Agent)(obs)

Take in an obs from environment and use agent.preprocessor to transform it into an internal state. Then use agent.learner to get the action. Return a pair of state => action.

source

Ju.DynaAgent — Type.

DynaAgent{Tl<:AbstractLearner, Tb<:AbstractTurnBuffer, Tm<:AbstractEnvironmentModel, Tpp<:Function} <: AbstractAgent
DynaAgent(learner::Tl, buffer::Tb, model::Tm, nsteps::Int=0, preprocessor::Tpp=identity, role=:anonymous) where {Tl, Tb, Tm, Tpp}

See more details at Section (8.2) on Page 162 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

See also: Agent, AbstractLearner, AbstractTurnBuffer, AbstractEnvironmentModel

source

Ju.DynaAgent — Method.

(agent::DynaAgent)(obs)

Take in an obs from environment and use agent.preprocessor to transform it into an internal state. Then use agent.learner to get the action. Return a pair of state => action.

source

Buffers

Ju.CircularSARDBuffer — Type.

CircularSARDBuffer is just an alias for CircularTurnBuffer{(:state, :action, :reward, :isdone)}.

source

Ju.CircularSARDBuffer — Method.

CircularSARDBuffer(capacity; state_type::Type=Int, action_type::Type=Int, state_size=(), action_size=())

capacity specifies how many latest turns the buffer will store at most.

Note

Note that, the length of state and action is 1 step longer in oder to store the state and action in the next step. This is the supposed behavior of SARD buffers.

source

Ju.CircularSARDSABuffer — Type.

CircularSARDSABuffer is just an alias for CircularTurnBuffer{(:state, :action, :reward, :isdone, :nextstate, :nextaction)}.

source

Ju.CircularSARDSBuffer — Type.

CircularSARDSBuffer is just an alias for CircularTurnBuffer{(:state, :action, :reward, :isdone, :nextstate)}.

source

Ju.CircularTurnBuffer — Type.

CircularTurnBuffer{names, types, Tbs} <: AbstractTurnBuffer{names, types}
CircularTurnBuffer{names, types}(capacities::NTuple{N, Int}, sizes::NTuple{N, NTuple{M, Int} where M}) where {names, types, N}

Using CircularArrayBuffer to store each element specified in names and types.. The memory of the buffer will be pre-allocated. In RL problems, three of the most common CircularArrayBuffer based buffers are: CircularSARDBuffer, CircularSARDSBuffer, CircularSARDSABuffer

source

Ju.EpisodeSARDBuffer — Type.

EpisodeSARDBuffer is just an alias for EpisodeTurnBuffer{(:state, :action, :reward, :isdone)}

source

Ju.EpisodeSARDSABuffer — Type.

EpisodeSARDSABuffer is just an alias for EpisodeTurnBuffer{(:state, :action, :reward, :isdone, :nextstate, :nextaction)}

source

Ju.EpisodeSARDSBuffer — Type.

EpisodeSARDSBuffer is just an alias for EpisodeTurnBuffer{(:state, :action, :reward, :isdone, :nextstate)}

source

Ju.EpisodeTurnBuffer — Type.

EpisodeTurnBuffer{names, types, Tbs} <: AbstractTurnBuffer{names, types}
EpisodeTurnBuffer{names, types}() where {names, types}

Using a Vector to store each element specified by names and types.

See also: EpisodeSARDBuffer, EpisodeSARDSBuffer, EpisodeSARDSABuffer

source

Learners

Ju.DifferentialTDLearner — Type.

DifferentialTDLearner(approximator::Tapp, π::Tp, α::Float64, β::Float64, R̄::Float64=0., n::Int=0, method::Symbol=:SARSA) where {Tapp<:AbstractApproximator, Tp<:PolicyOrSelector}= new{Tapp, Tp, method}(approximator, π, α, β, R̄, n)

See more details at Section (10.3) on Page 251 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.DoubleLearner — Type.

struct DoubleLearner{Tl <: OffPolicyTDLearner, Ts<:AbstractActionSelector} <: AbstractLearner
    Learner1::Tl
    Learner2::Tl
    selector::Ts
end

See more details at Section (6.7) on Page 126 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.GradientBanditLearner — Type.

struct GradientBanditLearner{Ts<:AbstractActionSelector, Tb<:Union{Float64, Function}} <: AbstractModelFreeLearner 
    Q::TabularQ 
    selector::Ts
    α::Float64
    baseline::Tb
end

See more details at Section (2.8) on Page 37 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.MonteCarloExploringStartLearner — Type.

MonteCarloExploringStartLearner(approximator::Tapp, π::Tp, π_start::RandomPolicy, γ::Float64, α::Float64 = 1.0; is_first_visit::Bool = true) where {Tapp,Tp}

See more details at Section (5.3) on Page 99 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.MonteCarloLearner — Type.

MonteCarloLearner(approximator::Tapp, π::Tp, γ::Float64=1., α::Float64 = 1.0, first_visit::Bool = true) where {Tapp,Tp}

See more details at Section (5.1) on Page 92 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.OffPolicyMonteCarloLearner — Type.

OffPolicyMonteCarloLearner(approximator::Tapp, π_behavior::Tpb, π_target::Tpt, γ::Float64, α::Float64 = 1.0; isfirstvisit::Bool = true, sampling::Symbol = :OrdinaryImportanceSampling,) where {Tapp,Tpb,Tpt}

See more details at Section (5.7) on Page 111 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.OffPolicyTDLearner — Type.

OffPolicyTDLearner(approximator::Tapp, π_behavior::Tpb, π_target::Tpt, γ::Float64, α::Float64, n::Int=0, method::Symbol=:SARSA_ImportanceSampling) where {Tapp<:AbstractApproximator, Tpb<:AbstractPolicy, Tpt<:AbstractPolicy}

See more details at Section (7.3) on Page 148 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.ReinforceBaselineLearner — Type.

mutable struct ReinforceBaselineLearner{Tapp<:AbstractApproximator, Tp<:AbstractPolicy}  <: AbstractModelFreeLearner 
    approximator::Tapp
    π::Tp
    αʷ::Float64
    αᶿ::Float64
    γ::Float64
end

See more details at Section (13.4) on Page 330 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.ReinforceLearner — Type.

struct ReinforceLearner{Tp<:AbstractPolicy}  <: AbstractModelFreeLearner 
    π::Tp
    α::Float64
    γ::Float64
end

See more details at Section (13.3) on Page 326 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.TDLearner — Type.

TDLearner(approximator::Tapp, π::Tp, γ::Float64, α::Float64, n::Int=0) where {Tapp<:AbstractVApproximator, Tp<:PolicyOrSelector} = new{Tapp, Tp, :SRS}(approximator, π, γ, α, n)
TDLearner(approximator::Tapp, π::Tp, γ::Float64, α::Float64, n::Int=0, method::Symbol=:SARSA) where {Tapp<:AbstractQApproximator, Tp<:PolicyOrSelector}

See more details at Section (7.1) on Page 142 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.TDλReturnLearner — Type.

struct TDλReturnLearner{Tapp <: AbstractApproximator, Tp <: PolicyOrSelector} <: AbstractModelFreeLearner 
    approximator::Tapp
    π::Tp
    γ::Float64
    α::Float64
    λ::Float64
end

See more details at Section (12.2) on Page 292 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Approximators

Value Functions

Ju.AggregationV — Type.

struct AggregationV{Tf<:Function} <: AbstractVApproximator{Int}
    table::Vector{Float64}
    f::Tf
end

Using a.f to map a state s into an Int, then use a.table to check the corresponding state value.

source

Ju.FourierV — Type.

FourierV <: AbstractVApproximator{Int}

struct FourierV <: AbstractVApproximator{Int}
    weights::Vector{Float64}
end

Using Fourier cosine basis to approximate the state value. weights is the featur vector.

See more details at Section (9.5.2) on Page 211 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.FourierV — Method.

FourierV(order::Int)

By specifying the order, feature vector will be initialized with 0.

source

Ju.LinearV — Type.

struct LinearV <: AbstractVApproximator{Int}
    features::Array{Float64, 2}
    weights::Vector{Float64}
end

Using a matrix features to represent each state along with a vector of weights.

See more details at Section (9.4) on Page 205 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.PolynomialV — Type.

struct PolynomialV <: AbstractVApproximator{Int}
    weights::Vector{Float64}
end

See more details at Section (9.5.1) on Page 210 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.TabularV — Type.

struct TabularV <: AbstractVApproximator{Int}
    table::Vector{Float64}
end

Using a table of type Vector{Float64} to record the state values.

source

Ju.TilingsV — Type.

TilingsV{Tt<:Tiling} <: AbstractVApproximator{Vector{Float64}}
TilingsV(tilings::Vector{Tt}) where Tt<:Tiling

Using a vector of tilings to encode state. Each tiling has an independent weight.

See more details at Section (9.5.4) on Page 217 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

See also: Tiling, TilingsQ

source

Action Value Functions (Q functions)

Ju.TabularQ — Type.

struct TabularQ <: AbstractQApproximator{Int, Int}
    table::Array{Float64, 2}
end

Using a table of type Array{Float64,2} to record the action value of each state.

source

Ju.TabularQ — Type.

TabularQ(ns::Int, na::Int=1, init::Float64=0.)

Initial a table of size (ns, na) filled with value of init.

source

Ju.TilingsQ — Type.

TilingsQ{Tt<:Tiling} <: AbstractQApproximator{Vector{Float64}, Int}
TilingsQ(tilings::Vector{Tt}, nactions) where Tt<:Tiling

The only difference compared to TilingsV is that now the weight of each tiling is a matrix.

source

Environment Models

Ju.DeterministicDistributionModel — Type.

struct DeterministicDistributionModel <: AbstractDistributionModel table::Array{Vector{NamedTuple{(:nextstate, :reward, :prob), Tuple{Int, Float64, Float64}}}, 2} end

Store all the transformations in the table.

source

Ju.DynamicDistributionModel — Type.

struct DynamicDistributionModel{Tf<:Function} <: AbstractDistributionModel
    f::Tf
    ns::Int
    na::Int
end

Using a general function f to store the transformations.

source

Ju.ExperienceSampleModel — Type.

  ExperienceSampleModel <: AbstractSampleModel

Generate a turn sample based on previous experiences.

source

Ju.PrioritizedSweepingSampleModel — Type.

PrioritizedSweepingSampleModel <: AbstractSampleModel
PrioritizedSweepingSampleModel(θ::Float64=1e-4)

See more details at Section (8.4) on Page 168 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Policies

Ju.DeterministicPolicy — Type.

struct DeterministicPolicy <: AbstractPolicy
    table::Vector{Int}
    nactions::Int
end

The action to be adopt is stored in table.

source

Ju.EpsilonGreedyPolicy — Type.

struct EpsilonGreedyPolicy <: AbstractPolicy
    table::Vector{Int}
    nactions::Int
    ϵ::Float64
end

The best actions are stored in the table. However the best action will only be taken at a portion of 1 - ϵ.

See also: EpsilonGreedySelector

source

Ju.LinearPolicy — Type.

struct LinearPolicy <: AbstractPolicy
    features::Array{Float64, 3}
    weights::Vector{Float64}
end

The probability of each action is calculate by features and weights and then normalized by softmax.

source

Ju.RandomPolicy — Type.

RandomPolicy(prob::Array{Float64, 2})
RandomPolicy(prob::Vector{Float64})

The probability of each action is predefined by prob. If prob is a vector, then all states share the same prob.

source

Action Selectors

Ju.AlternateSelector — Type.

AlternateSelector <: AbstractActionSelector

Used to ensure that all actions are selected alternatively.

AlternateSelector(n::Int)

n::Int means the optional actions are 1:n.

source

Ju.AlternateSelector — Method.

(s::AlternateSelector)(values::Any)

Ignore the action values, generate an action alternatively.

Example

julia> selector = AlternateSelector(3)
AlternateSelector(3, 0)

julia> any_state = 0 # for AlternateSelector, state can be anything

julia> [selector(any_state) for i in 1:10]  # iterate through all actions
10-element Array{Int64,1}:
 1
 2
 3
 1
 2
 3
 1
 2
 3
 1

source

Ju.EpsilonGreedySelector — Type.

EpsilonGreedySelector <: AbstractActionSelector
EpsilonGreedySelector(ϵ)

The best action is selected for a proportion $1 - \epsilon$ and a random action (with uniform probability) is selected for a proportion $\epsilon$.

Example

julia> selector = EpsilonGreedySelector(0.1)
EpsilonGreedySelector(0.1)

julia> countmap(selector([1,2,1,1]) for _ in 1:1000)
Dict{Any,Int64} with 4 entries:
  4 => 37
  2 => 915
  3 => 22
  1 => 26

source

Ju.EpsilonGreedySelector — Method.

(p::EpsilonGreedySelector)(values::AbstractArray)

Note

If multiple values with the same maximum value are found. Then a random one will be returned!

NaN will be filtered unless all the values are NaN. In that case, a random one will be returned.

source

Ju.UpperConfidenceBound — Type.

UpperConfidenceBound <: AbstractActionSelector
UpperConfidenceBound(na, c=2.0, t=0)

Arguments

na is the number of actions used to create a internal counter.

t is used to store current time step.

c is used to control the degree of exploration.

source

Ju.UpperConfidenceBound — Method.

(ucb::UpperConfidenceBound)(values::AbstractArray)

Unlike EpsilonGreedySelector, uncertaintyies are considered in UCB.

Note

If multiple values with the same maximum value are found. Then a random one will be returned!

\[A_t = \underset{a}{\arg \max} \left[ Q_t(a) + c \sqrt{\frac{\ln t}{N_t(a)}} \right]\]

See more details at Section (2.7) on Page 35 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.WeightedSample — Type.

WeightedSample <: AbstractActionSelector
WeightedSample()

source

Ju.WeightedSample — Method.

(p::WeightedSample)(values::AbstractArray)

Action values are normalized to have a sum of 1.0 and then used as the probability to sample a random action.

source