Interfaces

Interfaces

Several interfaces and abstract structures are carefully defined to make this package general enough to support both traditional reinforcement learning and deep reinforcement learning(although most algorithms in this package are targeting traditional RL currently).

Interfaces

Environment

Space

Agent

Buffer

Learner

Approximator

Environment Model

Policy

Action Selector

Environment

Ju.AbstractEnvironment — Type.

AbstractEnvironment{Tos, Tas, N}

Supertype for different kinds of environments with the observation space of type Tos and action space of type Tas. Here the N means the number of agents the environment can interact with.

See also: AbstractAsyncEnvironment, AbstractSyncEnvironment

source

Ju.AbstractSyncEnvironment — Type.

AbstractSyncEnvironment{Tos, Tas, N} <: AbstractEnvironment{Tos, Tas, N}

Supertype for synchronized environments with the observation space of type Tos and action space of type Tas. Here the N means the number of agents the environment can interact with. The Sync means that the environment will hang up and wait for the agent's input.

Required Methods	Brief Description

env(action)	Each environment env must be a functional object that receive an action as input and return a NamedTuple{(:observation, :reward, :isdone)}

observe(env)	Return a NamedTuple{(:observation, :isdone)}

reset!(env)	Reset the environment and return a NamedTuple{(:observation, :isdone)}

Optional Methods

observationspace(env)	Return the observation space of the environment. See also: AbstractSpace

actionspace(env)	Return the action space of the environment. See also: AbstractSpace

render(env)	Render the environment

source

Ju.AbstractAsyncEnvironment — Type.

AbstractAsyncEnvironment{Tos, Tas, N} <: AbstractEnvironment{Tos, Tas, N}

Opposed to AbstractSyncEnvironment, the Async environments will not wait for the agent's input. Instead it will keep running asynchronously.

source

Space

Ju.AbstractSpace — Type.

AbstractSpace

Supertype of AbstractContinuousSpace and AbstractDiscreteSpace.

source

Ju.AbstractContinuousSpace — Type.

AbstractContinuousSpace <: AbstractSpace

Supertype of different kinds of continuous spaces.

Required Methods	Brief Description

sample(space)	Get a random sample from the space

Base.in(space, x)	Test whether x is in the space

Optional Methods

Base.==(spaceX, spaceY)	Check whether two spaces are the same

Base.eltype(space)	Return the type of the sample in a space

See also: AbstractDiscreteSpace

source

Ju.AbstractDiscreteSpace — Type.

AbstractDiscreteSpace <: AbstractSpace

Supertype of different kinds of discrete spaces.

Required Methods	Brief Description

sample(space)	Get a random sample from the space

Base.in(space, x)	Test whether x is in the space

Base.size(space)	Return the size of the space in all dimensions

Optional Methods

Base.length(space)	Return the number of elements in that space. By default it will be *(Base.size(space)).

Base.==(spaceX, spaceY)	Check whether two spaces are the same

Base.eltype(space)	Return the type of the sample in a space

See also: AbstractContinuousSpace

source

Agent

Ju.AbstractAgent — Type.

AbstractAgent

Supertype of agents. Usually, an agent needs to contain at least an AbstractLearner and an AbstractTurnBuffer.

Required Methods	Brief Description

agent(obs)	agent, an instance of an AbstractAgent, must be a functional object to receive an observation as input and return a pair of state and action (s => a)

update!(agent, s, a, r, d, s′[, a′])	Update the agent after an interaction with environment

source

Buffer

Ju.AbstractTurnBuffer — Type.

AbstractTurnBuffer{names, types} <: AbstractArray{NamedTuple{names, types}, 1}

AbstractTurnBuffer is supertype of a collection of buffers to store the interactions between agents and environments. It is a subtype of AbstractArray{NamedTuple{names, types}, 1} where names specifies which fields are to store and types is the coresponding types of the names.

Required Methods	Brief Description

Base.push!(b::AbstractTurnBuffer{names, types}, s[, a, r, d, s′, a′])	Push a turn info into the buffer. According to different names and types of the buffer b, it may accept different number of arguments

isfull(b)	Check whether the buffer is full or not

capacity(b)	The maximum length of buffer

Base.length(b)	Return the length of buffer

Base.getindex(b::AbstractTurnBuffer{names, types})	Return a turn of type NamedTuple{names, types}

Base.empty!(b)	Reset the buffer

Optional Methods

Base.size(b)	Return (length(b),) by default

Base.isempty(b)	Check whether the buffer is empty or not. Return length(b) == 0 by default

Base.lastindex(b)	Return length(b) by default

source

Learner

Ju.AbstractLearner — Type.

AbstractLearner

Supertype for learners. A learner takes in a state and returns an action.

source

Approximator

Ju.AbstractApproximator — Type.

AbstractApproximator

AbstractApproximator is a supertype of a collection of approximators(including Tabular, Linear, DNN etc.)

source

Ju.AbstractVApproximator — Type.

AbstractVApproximator{Ts} <: AbstractApproximator

A supertype of state value approximators with the state of type Ts.

Required Methods	Brief Description

V(s)	V, an instance of AbstractVApproximator, must be a functional object which takes in state s and return the estimated value of that state

update!(V, s, e)	Update the value approximator of state s by a difference of e

source

Ju.AbstractQApproximator — Type.

AbstractQApproximator{Ts, Ta} <: AbstractApproximator

A supertype of action value approximators with the state of type Ts and action of type Ta.

Required Methods	Brief Description

Q(s, a)	Q, an instance of AbstractQApproximator must be a functional object which takes in state s and action a. The estimated value of that (state, action) will be returned.

Q(s)	Return the estimations of all actions at state s

update!(Q, s, a, e)	Update the approximator Q of state s and action a with difference e

Optional Methods

Q(s, Val(:max))	Return the maximum value of Q(s)

Q(s, Val(:argmax))	Return the argmax of Q(s) (randomly break ties)

update!(Q, s, errors)	Update the approximator Q of state s with the difference errors of each action

source

Environment Model

Ju.AbstractEnvironmentModel — Type.

AbstractEnvironmentModel

A supertype of environment models. An environment model is used to simulate the behavior of an environment.

source

Ju.AbstractSampleModel — Type.

AbstractSampleModel <: AbstractEnvironmentModel

An AbstractSampleModel can be used to generate a random turn info to simulate an environment.

Required Methods	Brief Description

sample(model)	Return a turn info (s, a, r, d, s′)

update!(model, buffer, learner)	Update the model given a buffer and a learner

source

Ju.AbstractDistributionModel — Type.

AbstractDistributionModel <: AbstractEnvironmentModel

An AbstractDistributionModel can return all the possible following up states, rewards, isdone info and corresponding probability.

Required Methods	Brief Description

sample(model, s, a)	Return a vector of (s′, r, d, p)

update!(model, buffer, learner)	Update the model given a buffer and a learner

getstates(model)	Get the number of states of the model

getactions(model)	Get the number of actions of the model

source

Policy

Ju.AbstractPolicy — Type.

AbstractPolicy

Supertype for policies. A policy takes in a state and returns an action.

source

Action Selector

Ju.AbstractActionSelector — Type.

AbstractActionSelector

A subtype of AbstractActionSelector is used to generate an action given the estimated valud of different actions.

Required Methods	Brief Description

selector(values)	selector, an instance of AbstractActionSelector, must be a callable object which takes in an estimation and returns an action

source