Utilities

Utilities

Utilities

Core

Callbasks

Decays

Iterators

Helper Functions

Others

Core

The most widely used functions while solving RL problems are listed below:

Ju.policy_evaluation! — Method.

policy_evaluation!(V::AbstractVApproximator, π::AbstractPolicy, Model::AbstractDistributionModel; γ::Float64=0.9, θ::Float64=1e-4)

See more details at Section (4.1) on Page 75 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.policy_improvement! — Method.

policy_improvement!(V::AbstractVApproximator, π::AbstractPolicy, Model::AbstractDistributionModel; γ::Float64=0.9)

See more details at Section (4.2) on Page 76 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.policy_iteration! — Method.

policy_iteration!(V::AbstractVApproximator, π::AbstractPolicy, Model::AbstractDistributionModel; γ::Float64=0.9, θ::Float64=1e-4, max_iter=typemax(Int))

See more details at Section (4.3) on Page 80 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Ju.train! — Method.

train!(env, agent; callbacks::Tuple{Vararg{<:Function}}=(stop_at_step(1),))

Train an agent in env. By default callbacks is (stop_at_step(1),). So it will only train one step forward.

source

Ju.train! — Method.

train!(env::AbstractSyncEnvironment{Tss, Tas, N} where {Tss, Tas},
       agents::Tuple{Vararg{<:Agent{<:AbstractLearner, <:SARDSBuffer}, N}};
       callbacks::Tuple{Vararg{<:Function}}=(stop_at_step(1),)) where N

TODO: Add an AgentManager struct to better organize agents.

For sync environments of mulit-agents, it becomes much more complicated compared to the single agent environments. Here is an implementation for one of the most common cases. Each agent take an action alternately. In every step, all agents may observe partial/complete information of the environment from their own perspective.

You may consider to overwrite this function according to the problem you want to solve.

source

Ju.value_iteration! — Method.

value_iteration!(V::AbstractVApproximator, Model::AbstractDistributionModel; γ::Float64=0.9, θ::Float64=1e-4, max_iter=typemax(Int))

See more details at Section (4.4) on Page 83 of the book Sutton, Richard S., and Andrew G. Barto. Reinforcement learning: An introduction. MIT press, 2018.

source

Callbasks

The callbacks are used in train!. Here are some of the predefined callbacks:

Ju.rewards_of_each_episode — Method.

A callback(closure) which will record the total reward of each episode. Only support single agent yet.

source

Ju.steps_per_episode — Method.

A callback(closure) which will record the length of each episode

source

Ju.stop_at_episode — Function.

stop_at_episode(n::Int, is_show_progress::Bool=true)

Return a function, which will return false after n episodes. isend(env) is used to check if it is the end of an episode. is_show_progress will control whether print the progress meter or not.

source

Ju.stop_at_step — Function.

stop_at_step(n::Int, is_show_progress::Bool=true)

Return a function, which will return false after been called n times. is_show_progress will control whether print the progress meter or not.

source

Ju.stop_when_done — Method.

Return false when encountered an end of an episode

source

Decays

Decays are used to efficiently calculate some variables which change based on time step.

Ju.cached_inverse_decay — Method.

cached_inverse_decay()

Example

julia> f = cached_inverse_decay();

julia> f(:a) # cached!
1.0

julia> f(:a) # call again
0.5

julia> f(:a) # call again
0.3333333333333333

julia> f(:b) # a new cache
1.0

source

Ju.cached_sample_avg — Method.

cached_sample_avg()

Example

julia> f = cached_sample_avg();

julia> f(:a, 3)  # cache :a
3.0

julia> f(:a, 5)  # calculate avg value of  :a (3 + 5) / 2
4.0

julia> f(:a, 8)  # calculate avg value of  :a (3 + 5 + 8) /3
5.333333333333333

julia> f(:b, 0)  # cache another value
0.0

source

Ju.incrementer — Method.

incrementer(;start=1, step=1)

Example

julia> f = incrementer();

julia> [f() for _ in 1:3]
3-element Array{Int64,1}:
 1
 2
 3

source

Ju.inverse_decay — Method.

inverse_decay()

Example

julia> f = inverse_decay()
#25 (generic function with 1 method)

julia> [f() for _ in 1:5]
5-element Array{Float64,1}:
 1.0
 0.5
 0.3333333333333333
 0.25
 0.2

source

Ju.multiplier — Method.

multiplier(;start=1.0, ratio=1.0)

Example

julia> f = multiplier(ratio=0.9);

julia> [f() for _ in 1:3]
3-element Array{Float64,1}:
 1.0
 0.9
 0.81

source

Ju.sample_avg — Method.

sample_avg()

Example

julia> f = sample_avg();

julia> f(2)
2.0

julia> f(3) # (2+3)/2
2.5

julia> f(5) # (2+3+5)/3
3.3333333333333335

source

Iterators

Some iterators are very helpful while implementing traditional RL algorithms.

Ju.IsFirstVisit — Type.

IsFirstVisit(itr)

Return an iterator which signifies whether each element in itr occurs for the first time.

Example

julia> s = [1,2,3,1,4,2,5];

julia> is_first_visit(s)
7-element Array{Bool,1}:
  true
  true
  true
 false
  true
 false
  true

julia> Iterators.reverse(IsFirstVisit(s))
Base.Iterators.Reverse{IsFirstVisit{Array{Int64,1}}}(IsFirstVisit{Array{Int64,1}}([1, 2, 3, 1, 4, 2, 5]))

julia> collect(Iterators.reverse(IsFirstVisit(s)))
7-element Array{Any,1}:
  true
 false
  true
 false
  true
  true
  true

Warning

Although IsFirstVisit supports Iterators.reverse, you should still take care of the memory usage. Internally we will walk through the itr first and calculate the count of each unique element.

source

Ju.Reductions — Type.

struct Reductions{T<:Union{NamedTuple{(:init,)}, NamedTuple{()}}, F, I}
    op::F
    itr::I
    kw::T
end

Construct an iterator to get all the intermediate values while calling reduce(op, itr;kw...)

source

Ju.reductions — Method.

reductions(op, iter; init)

Return an Iterator of the intermediate values of the reduction (as per reduce) of iter by op.

Note

You can not apply Iterators.reverse to Reductions (due to time complexity). If you really want to get the reversed Reductions, consider collect first and then call reverse!

Example

julia> reductions(+, [2,3,4])
3-element Array{Int64,1}:
 2
 5
 9

julia> reductions(+, [2,3,4], init=3)
Reductions{NamedTuple{(:init,),Tuple{Int64}},typeof(+),Array{Int64,1}}(+, [2, 3, 4], (init = 3,))

julia> collect(reductions(+, [2,3,4], init=3))
4-element Array{Any,1}:
  3
  5
  8
 12

source

StatsBase.countmap — Method.

extend the countmap in StatsBase to support general iterator

source

Helper Functions

Following are some commonly used functions.

Ju.deletefirst! — Method.

deletefirst!(A::Vector, element)

Find the first element in A and delete it. == is used to compare equality.

source

Ju.findallmax — Method.

findallmax(A::AbstractArray)

Like findmax, but all the indices of the maximum value are returned.

Warning

All elements of value NaN in A will be ignored, unless all elements are NaN. In that case, the returned maximum value will be NaN and the returned indices will be collect(1:length(A))

#Examples

julia> findallmax([-Inf, -Inf, -Inf])
(-Inf, [1, 2, 3])

julia> findallmax([Inf, Inf, Inf])
(Inf, [1, 2, 3])

julia> findallmax([Inf, 0, Inf])
(Inf, [1, 3])

julia> findallmax([0,1,2,1,2,1,0])
(2, [3, 5])

source

Ju.importance_weight — Method.

importance_weight(π, b, states, actions)

Calculate the importance weight between the target policy π and behavior policy b given states and actions.

source

Ju.onehot — Function.

onehot(n::Int, x::Int, t::Type=Int; isdense::Bool=true)

If isdense is false, a SparseArray is returned.

source

Ju.reverse_discounted_rewards — Method.

reverse_discounted_rewards(rewards, γ)

Given the rewards and discount ratio γ, the discounted reward at each time step in the reversed order is returned. The returned object is of type Reductions.

source

Ju.reverse_importance_weights — Method.

reverse_importance_weights(π, b, states, actions)

Calculate the importance weight at each time step in the reversed order between the target policy π and behavior policy b given states and actions.

The returned object is of type Reductions

source

Others

StatsBase.sample — Function.

sample(b::CircularSARDBuffer, batch_size::Int)

Sample a random batch of States, Actions, Rewards, isDone, nextStates, nextActions without replacement of batch_size.

source

Ju.CircularArrayBuffer — Type.

CircularArrayBuffer{E, T, N}

Using a N dimension Array to simulate a circular buffer of N-1 dimensional elements. Here E is the type of element and T is same with the eltype of E. Call eltype(b::CircularArrayBuffer{E,T,N}) will return E.

Examples

julia> b = CircularArrayBuffer{Float64}(2)
0-element CircularArrayBuffer{Float64,Float64,1}

julia> push!(b, rand())
1-element CircularArrayBuffer{Float64,Float64,1}:
 0.9709012378596478

julia> push!(b, rand())
2-element CircularArrayBuffer{Float64,Float64,1}:
 0.9709012378596478
 0.4510778027035365

julia> push!(b, rand())
2-element CircularArrayBuffer{Float64,Float64,1}:
 0.4510778027035365
 0.6774251288208646

julia> b = CircularArrayBuffer{Array{Float64,2}}(3, (3,3))
3×3×0 CircularArrayBuffer{Array{Float64,2},Float64,3}

julia> push!(b, randn(3,3))
3×3×1 CircularArrayBuffer{Array{Float64,2},Float64,3}:
[:, :, 1] =
 -0.548592   0.926179  -1.40998
 -0.0888621  0.177208   0.342665
  0.0925987  1.18531    0.962738

source

Ju.Tiling — Type.

Tiling(ranges::NTuple{N, Tr}) where {N, Tr}

Using a tuple of ranges to simulate a tiling.

Example

julia> t = Tiling((1:2:5, 10:5:20))
Tiling{2,StepRange{Int64,Int64}}((1:2:5, 10:5:20), [1 3; 2 4])

julia> Ju.encode(t, (2, 12))  # encode into an Int
1

julia> Ju.encode(t, (2, 18))

julia> t2 = t - (1, 3)  # shift a little to get a new Tiling
Tiling{2,StepRange{Int64,Int64}}((0:2:4, 7:5:17), [1 3; 2 4])3

See also: TilingsV, TilingsQ

source